pyproject.toml
[project]
name = "damn"
version = "0.1.0"
description = "Keep Talking and Nobody Explodes 单人通关手册"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyqt5>=5.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyautogui>=0.9.54",
    "keyboard>=0.13.5",
    "pillow>=10.0.0",
    "ultralytics>=8.0.0",
    "pyyaml>=6.0",
]

[tool.uv]
# Windows 平台需要特殊处理 PyQt5
constraint-dependencies = [
    "pyqt5-qt5==5.15.2",  # 固定支持 Windows 的版本
]
